requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
beautifulsoup4
dnspython
emojis
Faker
ffmpeg-python
gitpython
git+https://github.com/AsmSafone/SafoneAPI
gtts
gpytranslate
hachoir
heroku3
httpx[http2]
httpx==0.25.2
lyricsgenius==3.0.1
motor
numpy
pillow
psutil
pycountry
pydub
pyfiglet
py-tgcalls==0.9.7
pykeyboard
pyrogram==2.0.106
python-dotenv
pyyaml
pytz
qrcode
requests
search_engine_parser
speedtest-cli
spotipy
telegraph
tgcrypto
unidecode
yt-dlp==2025.4.30
youtube-search
youtube-search-python
wget
